+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 1.1.1 - 2019-10-28
+### Changed
+- Used `dyn Trait` syntax everywhere since it is supported by downcast-rs's
+  min-supported rust version (1.33).
+
 ## 1.1.0 - 2019-10-07
 ### Added
 - Support for downcasting `Rc<Trait>` and `Arc<Trait>`.